requirements.txt
streamlit==1.28.1
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
python-docx==0.8.11
reportlab==4.0.7
Pillow==10.1.0
openpyxl==3.11.0
